(LlamaIndex) Generate prompt using content only
See implementation of SimpleDirectoryReader for details:
During the ETL/index, several metadata such as file_name, file_path,
are stored in the index.  Despite some metadata by default are excluded
for llm generation, somehow SimpleDirectoryReader thinks file_path is
important and not excluded.

For the purpose of the demo, I just want to include the text content
as the prompt to LLM